refactor: change file and folder structer
