+++ src/components/ArrangeMovies/Arrange.module.css
@@ -1,0 +1,6 @@
+.arrange {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border-bottom: 1px solid #555555;
+}
