+++ src/components/ArrangeMovies/Filter/Filter.module.css
@@ -1,0 +1,7 @@
+.filter__genres {
+  display: flex;
+  gap: 2rem;
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
