+++ src/components/ArrangeMovies/Filter/Filter.tsx
@@ -1,0 +1,22 @@
+import classes from "./Filter.module.css";
+import FilterOption from "./FilterOption";
+
+const filters = [
+  { id: "a1", title: "All" },
+  { id: "b2", title: "Documentary" },
+  { id: "c3", title: "Comedy" },
+  { id: "d4", title: "Horror" },
+  { id: "e5", title: "Crime" },
+];
+
+const Filter: React.FC = () => {
+  return (
+    <ul className={classes["filter__genres"]}>
+      {filters?.map((filter) => (
+        <FilterOption key={filter.id} id={filter.id} title={filter.title} isActive={filter.title === "All"} />
+      ))}
+    </ul>
+  );
+};
+
+export default Filter;
