+++ src/components/ArrangeMovies/Filter/FilterOption.module.css
@@ -1,0 +1,12 @@
+.filter__genre{
+  background-color: transparent;
+  text-transform: uppercase;
+  color: white;
+  border: none;
+  padding: 1rem 0;
+}
+
+.filter__genre--active{
+  border-bottom: 1px solid #f65261;
+  margin-bottom: -1px;
+}
