+++ src/index.tsx
@@ -9,7 +9,6 @@ const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement)
 root.render(
   <React.StrictMode>
     <ErrorBoundary>
-
       <App />
     </ErrorBoundary>
   </React.StrictMode>
